include pepperflashplugin-nonfree on install
I run chromium on pop os and couldn't view a flash video. I tried installing browser-plugin-freshplayer-pepperflash on it's own and it didn't do anything. Installing the additional non free plugin seemed to do the trick. The advise at 
https://askubuntu.com/questions/836261/pepperflashplugin-nonfree-package-installation-fails-since-chrome-54-is-out-oct
is to not use this package at all and instead use adobe-flashplugin. I've got it working for now so will trade the resultant vulnerabilities for having it work. However, once I'm out of exam season, I'll revisit. 
Simon