Make sure install location is in path
For me this was /usr/lib/cuda/bin/ but I wasn't sure if different configurations may put it in different locations so I left that out. Not sure the wording on the change is the best, but something like this would have saved me some trouble with verifying the install.