updating software triage
Removed 'Causes.' There's too much negative verbiage in there and it's not necessary. You have symptoms and you have solutions.
Are all the symptoms necessary? Can they be described without negative words? 
Changed suspend-reboot-shutdown section
Adding suspend and reboot in the beginning doesn't make sense to me and might not to the customer. It overcomplicates the first suggestion to try turning it off and on again. 
Removed Solutions section: causes unnecessary scrolling and is unnecessary if there is a table of contents to the side. Delays getting to the actual solution. The verbiage was fun, but after more thought, it's unnecessary.
The purpose of this article is to get the customer to a solution when they are having software issues.