Update doc to mention install of ffmpeg
I ran through this setup this morning and was able to get howdy working. I just wanted to update the doc to include a mention of the need to install ffmpeg. When I originally tried to run ffplay I got an error saying that I should try installing ffmpeg first. Hope it helps, thanks!