Changed text for machine checks
I think that there is a typo in the memory section regarding corrected and uncorrected errors.  Customers without ECC RAM should never see corrected errors in their logs.  I also think you are intending to say that if you have ECC RAM and there are a large number of corrections in a short period of time it may still indicate a fault.