Add note on handling cryptdata entry issues
Upon restoring my bootloader after a failed motherboard and hardware replacements, I ran into the same issue as found here:

https://teddit.sethforprivacy.com/r/pop_os/comments/h8hl4x/systemd_boot_configuration_repair_broken_after/

Removing the extra characters after `cryptdata` remedied the issue and allowed me to boot properly again, so I figured it would be a good addition. If there is a less commonly used doc or better place for this, feel free to reject this PR and point me in the right direction!