Clarified DP-MST functionality
The paragraph title "A Note About Multi-Monitor Support" had some not quite accurate information. I reworded and elaborated on DP-MST, which is what powers the monitor daisy-chaining feature. But, this does not require a monitor to implement, and there are many MST hubs on the market that do the same thing. If any Cx are just looking for expanded video outputs and not a full on dock, an MST hub could be a better option for them.