Update troubleshooting commands for v22.04
Some of the instructions on this document no longer apply for Pop!OS v22.04, so I've split up those commands into a version for v22.04 and a verison for previous releases. I also added a note to the section about the `module-bluetooth-discover` Pulseaudio module, which is no longer used in v22.04.