Added note for using GNOME tweaks to restore theme settings
Great that there is a page detailing how to use another DE, however upon removal of the new DE and switching back to Pop_!OS default shell various visual elements need to be fixed manually with GNOME tweaks, which I found out the hard way. This adds a note mentioning these steps in the hope it helps others.