Adding in a line for draining caps.
"5. Disconnect the power cable and switch the button on the power supply to off. Hold down the on/off button on the computer three times for 10 seconds each. This will drain capacitors and can potentially fix a failure to power on."
This was not in the article previously. Helped someone yesterday who had gone through this article before calling. Draining caps got his system working and should probably be added into this article.