Update for Plugable UD-ULTC4K (#1064)
* Update for Plugable UD-ULTC4K

It does not natively support multiple monitors with the Lemur. I tested this myself and when it didn't work, I confirmed this with Pluggable's  support.

Apologies for not linking to the PR number, but I don't know what the number will be before the PR is made.

* fix indentation

* fix: Update URL to point to PR, tweak wording

* fix: Add specific generation of Lemur Pro

* replace parentheses with square brackets 

To match the style of other entries.

Co-authored-by: Randall White <N3M0-22@pm.me>
Co-authored-by: Jacob Kauffmann <jacobgkau@users.noreply.github.com>