Update active-directory-client.md
Without ad_gpo_access_control = permissive in the [domain] section of the SSSD conf I could not log in to my Server 2022 AD from PopOS. I've found dozens of places on the internet where people have complained that they can not log in, just like the issue I had. None of those people seemed to find the fix. If this isn't added we should at least leave it as a comment for those who can't log in on their AD instance.